# qiyu-design/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "栖屿设计 API",
    "description": "栖屿设计 · 室内设计企业网站前后端接口（FastAPI）",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "name": "cat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(private|small|apartment)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cat"
            }
          },
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "风格标签，逗号分隔多值",
              "title": "Style"
            },
            "description": "风格标签，逗号分隔多值"
          },
          {
            "name": "house_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "户型标签，逗号分隔多值",
              "title": "House Type"
            },
            "description": "户型标签，逗号分隔多值"
          },
          {
            "name": "area_range",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Range"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(latest|hottest)$",
              "default": "latest",
              "title": "Sort"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ApiListResponse_CaseListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Case Detail",
        "operationId": "case_detail_api_v1_cases__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CaseDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{slug}/view": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Bump Case View",
        "description": "浏览量上报：IP+slug 60s 窗口去重（文档 §6.3.4 限流：同 IP 同 slug 60s 计 1 次）。",
        "operationId": "bump_case_view_api_v1_cases__slug__view_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/packages": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List Packages",
        "operationId": "list_packages_api_v1_packages_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(single_space|whole_house|style)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ApiListResponse_PackageListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/packages/{slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Package Detail",
        "operationId": "package_detail_api_v1_packages__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PackageDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/news": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List News",
        "operationId": "list_news_api_v1_news_get",
        "parameters": [
          {
            "name": "cat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(company|industry)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cat"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ApiListResponse_NewsOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/{slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "News Detail",
        "operationId": "news_detail_api_v1_news__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NewsOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/careers": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List Careers",
        "operationId": "list_careers_api_v1_careers_get",
        "parameters": [
          {
            "name": "cat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(social|campus)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_CareerOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiListResponse_CaseListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "pageSize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagesize"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ApiListResponse[CaseListItem]"
      },
...
      "ApiListResponse_PackageListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PackageListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "pageSize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagesize"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ApiListResponse[PackageListItem]"
      },
...
      "ApiResponse_ApiListResponse_CaseListItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiListResponse_CaseListItem_"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[ApiListResponse[CaseListItem]]"
      },
...
      "ApiResponse_ApiListResponse_NewsOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiListResponse_NewsOut_"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[ApiListResponse[NewsOut]]"
      },
      "ApiResponse_ApiListResponse_PackageListItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiListResponse_PackageListItem_"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[ApiListResponse[PackageListItem]]"
      },
...
      "ApiResponse_CaseDetailOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDetailOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[CaseDetailOut]"
      },
...
      "ApiResponse_NewsOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[NewsOut]"
      },
      "ApiResponse_PackageDetailOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PackageDetailOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[PackageDetailOut]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
      "ApiResponse_list_CareerOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CareerOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[CareerOut]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PackageDetailOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "cover": {
            "type": "string",
            "title": "Cover"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "applicable_house_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable House Type"
          },
          "price_per_sqm": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price Per Sqm",
            "default": 0
          },
          "price_from": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price From",
            "default": 0
          },
          "area_step_coefficient": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Area Step Coefficient",
            "default": 1.0
          },
          "price_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Note"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "process_steps": {
            "items": {
              "$ref": "#/components/schemas/PackageProcessStepOut"
            },
            "type": "array",
            "title": "Process Steps"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "type",
          "cover",
          "status"
        ],
        "title": "PackageDetailOut",
        "description": "详情：含 description 长文 + 流程步骤。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}